pyproject.toml
[build-system]
requires = ["poetry-core>=1.2.0", "wheel"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "agent-crab"
version = "0.0.1"
description = "Cross-platfrom Agent Benchmark for Multimodal Embodied Language Model Agents"

authors = ["Tianqi Xu <tianqi.xu@kaust.edu.sa>"]
maintainers = ["Tianqi Xu <tianqi.xu@kaust.edu.sa>"]

packages = [{ include = "crab" }]

[tool.poetry.dependencies]
python = "^3.10"

# core
docstring-parser = "^0"
networkx = "^3"
dill = "^0.3.8"
pydantic = "^2.6"
lxml = "^5.2.2"

# desktop actions
pillow = "^10.2.0"
mss = "^9.0.1"
psutil = "^5.9.8"
pyautogui = "^0.9.3"
pyperclip = "^1.8.2"

# environment
python-vagrant = "^1.0.0"


# client
httpx = "*"

# agent
openai = "^1.12.0"
google-generativeai = "^0.5.2"
anthropic = "^0.26.0"
groq = "^0.5.0"
ollama = "^0.2.0"

# text ocr
pytesseract = { version = "^0.3.10", optional = true }
easyocr = "^1.7.1"

# groundingdino object detection
torch = { version = "^2.2.1", optional = true }
torchvision = { version = "^0.17.1", optional = true }
numpy = { version = "^1.26.4", optional = true }
opencv-python = { version = "^4.9.0.80", optional = true }
transformers = { version = "^4.38.2", optional = true }
addict = { version = "^2.4.0", optional = true }
yapf = { version = "^0.40.2", optional = true }
matplotlib = { version = "^3.8.3", optional = true }
pycocotools = { version = "^2.0.7", optional = true }
timm = { version = "^0.9.16", optional = true }

# server
fastapi = { extras = ["all"], version = "0.109.1", optional = true }
pydantic-settings = { version = "^2", optional = true }
uvicorn = { extras = ["standard"], version = "^0.27.0.post1", optional = true }

# radar plot
plotly = { version = "^5.20.0", optional = true }

# types
types-pyautogui = "^0.9.3.20240106"
types-psutil = "^5.9.5.20240205"
types-networkx = "^3.2.1.20240210"
pyexcel-ods = "^0.6.0"
odfpy = "^1.4.1"
beautifulsoup4 = "^4.12.3"
termcolor = "^2.4.0"

[tool.poetry.extras]
server = ["fastapi", "pydantic-settings", "uvicorn"]
visual-prompt = [
    "pytesseract",
    "torch",
    "torchvision",
    "numpy",
    "opencv-python",
    "transformers",
    "addict",
    "yapf",
    "matplotlib",
    "pycocotools",
    "timm",
]

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"
pytest = "^8.0.0"
ruff = "^0.1.14"
ipykernel = "^6.29.3"
pandas = "^1"

[tool.ruff]
select = ["E501", "E4", "E7", "E9", "F", "I"]
ignore = ["E731"]
ignore-init-module-imports = true
